refactor: Extract plugin and export stats types
